requirements.txt
streamlit==1.32.2
pandas==2.2.1
plotly==5.20.0
xgboost==2.0.3
prophet==1.1.5
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
requests==2.31.0
scipy==1.12.0
ollama>=0.6.1
pymssql==2.3.0
sqlalchemy==2.0.28
